replacing code to different directory
