+++ e2e-test/src/test/java/CeUsersTests/CeUsersTest.java
@@ -1,7 +1,6 @@
 package CeUsersTests;
 
 import CeUsersTests.controller.CeUsersEndpoints;
-import CeUsersTests.domain.CEUsers;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.restassured.response.Response;
@@ -11,6 +10,7 @@ import org.testcontainers.containers.DockerComposeContainer;
 import org.testcontainers.containers.wait.strategy.Wait;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
+import usersapplication.domain.Users;
 
 import java.io.File;
 
@@ -37,19 +37,19 @@ public class CeUsersTest {
     @Test
     public void CrudActions() throws JsonProcessingException {
         // Arrange
-        CEUsers ceUsers1 = CEUsers.builder().id(1L).isActive(true).firstName("Harry")
-                .lastName("Wit, de").address("Straat 2").occupation("TAE").build();
-        CEUsers ceUsers2 = CEUsers.builder().id(2L).isActive(true).firstName("Jos")
-                .lastName("Lelijk").address("Weg 400").occupation("TAE").build();
-        CEUsers ceUsers3 = CEUsers.builder().id(3L).isActive(false).firstName("Jos")
-                .lastName("Lelijk").address("Weg 400").occupation("TAE").build();
-        CEUsers updatedCeUsers1 = CEUsers.builder().id(1L).isActive(true).firstName("Johan")
-                .lastName("Rode, de").address("Steegje 22").occupation("FAB").build();
+        Users users1 = Users.builder().id(1L).isActive(true).firstName("Harry")
+                .lastName("Wit, de").address("Straat 2").occupation("TAE").workingConditionsId(1).build();
+        Users users2 = Users.builder().id(2L).isActive(true).firstName("Jos")
+                .lastName("Lelijk").address("Weg 400").occupation("TAE").workingConditionsId(1).build();
+        Users users3 = Users.builder().id(3L).isActive(false).firstName("Jos")
+                .lastName("Lelijk").address("Weg 400").occupation("TAE").workingConditionsId(1).build();
+        Users updatedUsers1 = Users.builder().id(1L).isActive(true).firstName("Johan")
+                .lastName("Rode, de").address("Steegje 22").occupation("FAB").workingConditionsId(1).build();
 
         // Act - POST
-        Response postResponse1 = CeUsersEndpoints.postNewCeUser(objectMapper.writeValueAsString(ceUsers1));
-        Response postResponse2 = CeUsersEndpoints.postNewCeUser(objectMapper.writeValueAsString(ceUsers2));
-        Response postResponse3 = CeUsersEndpoints.postNewCeUser(objectMapper.writeValueAsString(ceUsers3));
+        Response postResponse1 = CeUsersEndpoints.postNewCeUser(objectMapper.writeValueAsString(users1));
+        Response postResponse2 = CeUsersEndpoints.postNewCeUser(objectMapper.writeValueAsString(users2));
+        Response postResponse3 = CeUsersEndpoints.postNewCeUser(objectMapper.writeValueAsString(users3));
 
         // Assert - POST
         assertThat(postResponse1.statusCode()).isEqualTo(HttpStatus.SC_CREATED);
@@ -59,37 +59,37 @@ public class CeUsersTest {
         // Act - GET
         Response getAllActiveResponse = CeUsersEndpoints.getAllActive();
         Response getAllInactiveResponse = CeUsersEndpoints.getAllInactive();
-        Response getResponse1 = CeUsersEndpoints.getById(ceUsers1.getId().toString());
-        Response getResponse2 = CeUsersEndpoints.getById(ceUsers2.getId().toString());
-        Response getResponse3 = CeUsersEndpoints.getById(ceUsers3.getId().toString());
+        Response getResponse1 = CeUsersEndpoints.getById(users1.getId().toString());
+        Response getResponse2 = CeUsersEndpoints.getById(users2.getId().toString());
+        Response getResponse3 = CeUsersEndpoints.getById(users3.getId().toString());
 
         // Assert - GET
         assertThat(getAllActiveResponse.statusCode()).isEqualTo(HttpStatus.SC_OK);
         assertThat(getAllActiveResponse.body().print()).isNotEqualTo("[" +
-                objectMapper.writeValueAsString(ceUsers1) + "," +
-                objectMapper.writeValueAsString(ceUsers2) + "," +
-                objectMapper.writeValueAsString(ceUsers3 + "]"));
+                objectMapper.writeValueAsString(users1) + "," +
+                objectMapper.writeValueAsString(users2) + "," +
+                objectMapper.writeValueAsString(users3 + "]"));
         assertThat(getAllActiveResponse.body().print()).isEqualTo("[" +
-                objectMapper.writeValueAsString(ceUsers1) + "," +
-                objectMapper.writeValueAsString(ceUsers2) + "]");
+                objectMapper.writeValueAsString(users1) + "," +
+                objectMapper.writeValueAsString(users2) + "]");
         assertThat(getAllInactiveResponse.statusCode()).isEqualTo(HttpStatus.SC_OK);
         assertThat(getAllInactiveResponse.body().print()).isEqualTo("[" +
-                objectMapper.writeValueAsString(ceUsers3) + "]");
+                objectMapper.writeValueAsString(users3) + "]");
         assertThat(getResponse1.statusCode()).isEqualTo(HttpStatus.SC_OK);
-        assertThat(getResponse1.body().print()).isEqualTo(objectMapper.writeValueAsString(ceUsers1));
+        assertThat(getResponse1.body().print()).isEqualTo(objectMapper.writeValueAsString(users1));
         assertThat(getResponse2.statusCode()).isEqualTo(HttpStatus.SC_OK);
-        assertThat(getResponse2.body().print()).isEqualTo(objectMapper.writeValueAsString(ceUsers2));
+        assertThat(getResponse2.body().print()).isEqualTo(objectMapper.writeValueAsString(users2));
         assertThat(getResponse3.statusCode()).isEqualTo(HttpStatus.SC_OK);
-        assertThat(getResponse3.body().print()).isEqualTo(objectMapper.writeValueAsString(ceUsers3));
+        assertThat(getResponse3.body().print()).isEqualTo(objectMapper.writeValueAsString(users3));
 
         // Act - PUT & GET (by id)
-        Response putResponse = CeUsersEndpoints.putCeUser("1", objectMapper.writeValueAsString(updatedCeUsers1));
-        Response getUpdatedResponse1 = CeUsersEndpoints.getById(ceUsers1.getId().toString());
+        Response putResponse = CeUsersEndpoints.putCeUser("1", objectMapper.writeValueAsString(updatedUsers1));
+        Response getUpdatedResponse1 = CeUsersEndpoints.getById(users1.getId().toString());
 
         // Assert - PUT & GET (by id)
         assertThat(putResponse.statusCode()).isEqualTo(HttpStatus.SC_OK);
         assertThat(getUpdatedResponse1.statusCode()).isEqualTo(HttpStatus.SC_OK);
-        assertThat(getUpdatedResponse1.body().print()).isEqualTo(objectMapper.writeValueAsString(updatedCeUsers1));
+        assertThat(getUpdatedResponse1.body().print()).isEqualTo(objectMapper.writeValueAsString(updatedUsers1));
 
         // Act - DELETE & GET
         Response deleteResponse1 = CeUsersEndpoints.deleteCeUser("1");
