--- e2e-test/src/test/java/CeUsersTests/domain/CEUsers.java
@@ -1,19 +1,0 @@
-package CeUsersTests.domain;
-
-import lombok.*;
-
-@Builder
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class CEUsers {
-
-    private Long id;
-    private boolean isActive;
-    private String firstName;
-    private String lastName;
-    private String address;
-    private String occupation;
-
-}
