--- working-conditions-application/src/main/java/nl/conclusionexperts/workingconditionapplication/CeWorkingConditionsApplication.java
@@ -1,13 +1,0 @@
-package nl.conclusionexperts.workingconditionapplication;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CeWorkingConditionsApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CeWorkingConditionsApplication.class, args);
-    }
-
-}
