--- working-conditions-application/src/main/java/nl/conclusionexperts/workingconditionapplication/domain/NewUpdateWorkingConditions.java
@@ -1,21 +1,0 @@
-package nl.conclusionexperts.workingconditionapplication.domain;
-
-import lombok.*;
-import nl.conclusionexperts.workingconditionapplication.enums.CompanyLaptopTypes;
-import nl.conclusionexperts.workingconditionapplication.enums.SalaryGroups;
-
-import javax.persistence.*;
-
-@Builder
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class NewUpdateWorkingConditions {
-
-    private @Enumerated(EnumType.STRING) @Column(nullable = false) SalaryGroups salaryGroup;
-    private @Column(nullable = false) boolean companyCar;
-    private @Enumerated(EnumType.STRING) @Column(nullable = false)
-    CompanyLaptopTypes companyLaptop;
-
-}
