--- working-conditions-application/src/main/java/nl/conclusionexperts/workingconditionapplication/domain/salary/UpdateSalaryGroup.java
@@ -1,17 +1,0 @@
-package nl.conclusionexperts.workingconditionapplication.domain.salary;
-
-import lombok.*;
-
-import javax.persistence.Column;
-
-@Builder
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UpdateSalaryGroup {
-
-    private @Column(nullable = false) double minAmount;
-    private @Column(nullable = false) double maxAmount;
-
-}
