--- working-conditions-application/src/main/java/nl/conclusionexperts/workingconditionapplication/enums/CompanyLaptopTypes.java
@@ -1,5 +1,0 @@
-package nl.conclusionexperts.workingconditionapplication.enums;
-
-public enum CompanyLaptopTypes {
-    Type1, Type2, Type3, Type4, Type5
-}
