--- working-conditions-application/src/main/java/nl/conclusionexperts/workingconditionapplication/enums/OperatingSystem.java
@@ -1,5 +1,0 @@
-package nl.conclusionexperts.workingconditionapplication.enums;
-
-public enum OperatingSystem {
-    Windows, MacOSX, Linux, Solaris, BSD, Unix, Unknown;
-}
