--- working-conditions-application/src/main/java/nl/conclusionexperts/workingconditionapplication/enums/SalaryGroups.java
@@ -1,5 +1,0 @@
-package nl.conclusionexperts.workingconditionapplication.enums;
-
-public enum SalaryGroups {
-    A, B, C, D, E, F, G
-}
