--- working-conditions-application/src/main/java/nl/conclusionexperts/workingconditionapplication/repository/CEWorkingConditionsRepository.java
@@ -1,8 +1,0 @@
-package nl.conclusionexperts.workingconditionapplication.repository;
-
-import nl.conclusionexperts.workingconditionapplication.domain.WorkingConditions;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CEWorkingConditionsRepository extends JpaRepository<WorkingConditions, Long> { }
