+++ working-conditions-application/src/main/java/workingconditionapplication/WorkingConditionsApplication.java
@@ -1,0 +1,13 @@
+package workingconditionapplication;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WorkingConditionsApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(WorkingConditionsApplication.class, args);
+    }
+
+}
