+++ working-conditions-application/src/main/java/workingconditionapplication/domain/salary/UpdateSalaryGroup.java
@@ -1,0 +1,17 @@
+package workingconditionapplication.domain.salary;
+
+import lombok.*;
+
+import javax.persistence.Column;
+
+@Builder
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class UpdateSalaryGroup {
+
+    private @Column(nullable = false) double minAmount;
+    private @Column(nullable = false) double maxAmount;
+
+}
