+++ working-conditions-application/src/main/java/workingconditionapplication/enums/CompanyLaptopTypes.java
@@ -1,0 +1,5 @@
+package workingconditionapplication.enums;
+
+public enum CompanyLaptopTypes {
+    Type1, Type2, Type3, Type4, Type5
+}
