+++ working-conditions-application/src/main/java/workingconditionapplication/enums/OperatingSystem.java
@@ -1,0 +1,5 @@
+package workingconditionapplication.enums;
+
+public enum OperatingSystem {
+    Windows, MacOSX, Linux, Solaris, BSD, Unix, Unknown;
+}
