+++ working-conditions-application/src/main/java/workingconditionapplication/enums/SalaryGroups.java
@@ -1,0 +1,5 @@
+package workingconditionapplication.enums;
+
+public enum SalaryGroups {
+    A, B, C, D, E, F, G
+}
