+++ working-conditions-application/src/main/java/workingconditionapplication/repository/CEWorkingConditionsRepository.java
@@ -1,0 +1,8 @@
+package workingconditionapplication.repository;
+
+import workingconditionapplication.domain.WorkingConditions;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CEWorkingConditionsRepository extends JpaRepository<WorkingConditions, Long> { }
