+++ working-conditions-application/src/main/java/workingconditionapplication/repository/salary/SalaryGroupRepository.java
@@ -1,0 +1,13 @@
+package workingconditionapplication.repository.salary;
+
+import workingconditionapplication.domain.salary.SalaryGroup;
+import workingconditionapplication.enums.SalaryGroups;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SalaryGroupRepository extends JpaRepository<SalaryGroup, Long> {
+
+    SalaryGroup findUserBySalaryGroupCode(SalaryGroups salaryGroupCode);
+
+}
