--- working-conditions-application/src/test/java/nl/conclusionexperts/workingconditionapplication/CeWorkingConditionsApplicationTests.java
@@ -1,13 +1,0 @@
-package nl.conclusionexperts.workingconditionapplication;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CeWorkingConditionsApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
