+++ working-conditions-application/src/test/java/workingconditionapplication/WorkingConditionsApplicationTests.java
@@ -1,0 +1,13 @@
+package workingconditionapplication;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class WorkingConditionsApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
